+++ yash-cli/tests/scripted_test/false-p.sh
@@ -3,6 +3,5 @@
 posix="true"
 
 test_OE -e n 'false'
-PATH=
 false
 __IN__